Don't throw error in reporter if coverage was not run
If tests weren't run with instrumented code `__coverage__` never gets
defined in the global space. The previous check would throw
`Uncaught ReferenceError: __coverage__ is not defined`. I've updated
the check for determining whether or not to send the coverage command
to phantom to account for the case where `__coverage__` is not defined.