Modified template.js to copy reporter.js to the correct path.
fixes #31 :

If Jasmine's options.outfile is changed then coverage doesn't run.
The SpecRunner template seems to be generated correctly,
but grunt-template-jasmine-istanbul/reporter.js is not being copied into the
 .grunt/grunt-contrib-jasmine directory.

This commit modifies the exports.process method of template.js in order to
use the original path when copying reporter.js.